Better attempt at automatically handling the Still There? popup.
Wait for the Continue button in the Still There? popup to appear and
click it to keep the stream playing.  This version works better than
the previous one.  It creates a MutationObserver to wait for changes
to the HTML and only checks for the Continue button in the changed
subtrees when they occur.  Consequently, there is no polling and in
testing so far, the popup is dismissed before it is even visible.